move timer to layout, rename scene components
