+++ frontend/src/components/header.tsx
@@ -4,6 +4,7 @@ import { ModalMenu } from "./ui/modal-menu";
 import { RoomStage } from "@/state/features/room";
 import { GamePhase } from "@/state/features/game";
 import { cn } from "@/lib/utils";
+import { Timer } from "./ui/timer";
 
 const hideHeader = (stage: RoomStage, phase: GamePhase) => {
 	return phase === GamePhase.Drawing || stage === RoomStage.PreGame;
@@ -13,6 +14,12 @@ export function UIHeader() {
 	const stage = useSelector((state: RootState) => state.room.stage);
 	const phase = useSelector((state: RootState) => state.game.phase);
 
+	const deadline = useSelector(
+		(state: RootState) => state.game.currentPhaseDeadline
+	);
+
+	const showTimer = stage === RoomStage.Playing;
+
 	return (
 		<header
 			className={cn(
@@ -20,6 +27,7 @@ export function UIHeader() {
 				hideHeader(stage, phase) ? "hidden sm:flex" : "flex"
 			)}
 		>
+			{showTimer && <Timer endTime={deadline} />}
 			<div className="ml-auto">
 				<ModalMenu />
 			</div>
