+++ frontend/src/components/views/playing/drawing/drawing-drawer-view.tsx
@@ -4,21 +4,18 @@ import Canvas from "@/components/canvas";
 import { CanvasTools, ColorSliders } from "@/components/canvas-tools";
 import { GameRole } from "@/state/features/game";
 import { Guesses } from "./guesses";
-import { Timer } from "@/components/ui/timer";
-import { HillScene } from "@/components/scenes/hill-scene";
+import { SkyScene } from "@/components/scenes/sky-scene";
 import { BobbingDoodle } from "@/components/doodle/bobbing-doodle";
 import { AnimatePresence } from "framer-motion";
 import { AirplaneDoodle } from "@/components/doodle/airplane-doodle";
+import { ModalMenu } from "@/components/ui/modal-menu";
 export function DrawingDrawerView() {
-	const deadline = useSelector(
-		(state: RootState) => state.game.currentPhaseDeadline
-	);
 	const selectedWord = useSelector(
 		(state: RootState) => state.game.selectedWord
 	);
 
 	return (
-		<HillScene>
+		<SkyScene>
 			<div className="mx-auto mb-auto xl:my-auto flex flex-col lg:gap-2 gap-1  relative z-50">
 				<div className="flex w-full h-full xl:items-start items-center justify-center lg:gap-4 flex-col xl:flex-row relative">
 					<div className="flex items-center justify-center gap-6">
@@ -27,13 +24,15 @@ export function DrawingDrawerView() {
 						</div>
 						<div className="flex flex-col items-center justify-center max-w-[800px] w-screen lg:w-auto">
 							<div className="flex justify-between w-full items-center lg:items-end py-2 px-2">
-								<div className="lg:text-2xl">
+								<div className="lg:text-2xl flex items-end gap-2">
 									You're drawing:{" "}
 									<span className="text-lg lg:text-2xl font-bold">
 										{selectedWord}
 									</span>
 								</div>
-								<Timer endTime={deadline} playSound />
+								<div className="lg:hidden">
+									<ModalMenu />
+								</div>
 							</div>
 							<Canvas
 								padding={10}
@@ -68,6 +67,6 @@ export function DrawingDrawerView() {
 				animateTo={{ left: "85%", top: "55%", rotate: 30, opacity: 1 }}
 				leaveTo={{ left: "105%", top: "55%", rotate: 20 }}
 			/>
-		</HillScene>
+		</SkyScene>
 	);
 }
