+++ frontend/src/components/views/playing/picking/picking-guesser-view.tsx
@@ -2,8 +2,7 @@ import { generateAvatar } from "@/lib/avatar";
 import { useSelector } from "react-redux";
 import { RootState } from "@/state/store";
 import { getPickingPlayer } from "@/lib/player";
-import { Timer } from "@/components/ui/timer";
-import { HillScene } from "@/components/scenes/hill-scene";
+import { SkyScene } from "@/components/scenes/sky-scene";
 import { BobbingDoodle } from "@/components/doodle/bobbing-doodle";
 import { AnimatePresence } from "framer-motion";
 import { AirplaneDoodle } from "@/components/doodle/airplane-doodle";
@@ -11,9 +10,6 @@ import { AirplaneDoodle } from "@/components/doodle/airplane-doodle";
 export function PickingGuesserView() {
 	const players = useSelector((state: RootState) => state.room.players);
 	const pickingPlayer = getPickingPlayer(players);
-	const deadline = useSelector(
-		(state: RootState) => state.game.currentPhaseDeadline
-	);
 	const isFirstPhase = useSelector(
 		(state: RootState) => state.game.isFirstPhase
 	);
@@ -21,10 +17,7 @@ export function PickingGuesserView() {
 	const avatarSvg = generateAvatar(pickingPlayer.avatarSeed);
 
 	return (
-		<HillScene>
-			<div className="absolute lg:top-10 lg:right-10 top-4 right-4">
-				<Timer endTime={deadline} />
-			</div>
+		<SkyScene>
 			<img src={avatarSvg} className="w-20 h-20 rounded-lg shadow-accent" />
 			<h1 className="text-2xl lg:text-3xl font-bold px-8 lg:px-0 text-center">
 				{pickingPlayer.name} is picking a word
@@ -61,6 +54,6 @@ export function PickingGuesserView() {
 				animateTo={{ left: "45%", top: "65%", rotate: 30, opacity: 1 }}
 				leaveTo={{ left: "185%", top: "55%", rotate: 30 }}
 			/>
-		</HillScene>
+		</SkyScene>
 	);
 }
